+++ examples/example_riverpod/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:modularity_flutter/modularity_flutter.dart';
 import 'src/riverpod_module.dart';
 
+final _observer = RouteObserver<ModalRoute<dynamic>>();
+
 void main() {
   runApp(const MyApp());
 }
@@ -12,9 +14,10 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ModularityRoot(
+      observer: _observer,
       child: MaterialApp(
         title: 'Riverpod Example',
-        navigatorObservers: [Modularity.observer],
+        navigatorObservers: [_observer],
         home: ModuleScope(module: CounterModule(), child: const CounterPage()),
       ),
     );